Day 25.
This is more bits about graphs than I previously understood.  I'd
looked around in Data.Graph in the past, but never got a good
understanding for how the bits piece together since I don't work with
graphs enough, but this was pretty straightforward once I knew what to
look for.